time_trigger/requirements.txt
adal==1.2.7
appdirs==1.4.4
argon2-cffi 
attrs
azure-applicationinsights==0.1.0
azure-batch==4.1.3
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.27
azure-core==1.13.0
azure-cosmos
azure-datalake-store==0.0.52
azure-eventgrid==1.3.0
azure-graphrbac==0.40.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.0
azure-mgmt==4.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-cdn==3.1.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==4.6.2
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-containerservice==4.4.0
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==1.0.0
azure-mgmt-eventhub==2.6.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iotcentral==0.1.0
azure-mgmt-iothub==0.5.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.1
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.0
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.7.0
azure-mgmt-notificationhubs==2.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.2.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.1.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
azure-servicebus==7.1.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.7
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
backcall
bleach 
certifi==2020.12.5
cffi
chardet==4.0.0
click 
cycler==0.10.0
decorator 
defusedxml 
entrypoints==0.3
et-xmlfile==1.1.0
Flask
idna==2.10
importlib-metadata==2.1.1
ipykernel
ipython
ipython-genutils 
ipywidgets 
isodate==0.6.0
itsdangerous 
jedi==0.17.0
Jinja2 
joblib==1.0.1
jsonschema 
jupyter==1.0.0
jupyter-client
jupyter-console
jupyter-core
jupyterlab-widgets
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
msrest==0.6.21
msrestazure==0.6.4
nbconvert==5.6.1
nbformat 
notebook
numpy
openpyxl==3.0.7
packaging 
Pillow==8.2.0
py4j==0.10.9
pyee==8.1.0
PyJWT==2.0.1
pyppeteer==0.2.5
python-dateutil
pyzmq==20.0.0
qtconsole
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
terminado==0.9.4
threadpoolctl==2.1.0
tqdm==4.61.0
traitlets==4.3.3
uamqp==1.3.0
urllib3==1.26.4
webencodings==0.5.1
websockets==8.1
widgetsnbextension==3.5.1